+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - `\Wizaplace\Favorite\FavoriteService::addDeclinationToUserFavorites` is now taking a `string $declinationId` parameter instead of a `int $declinationId` one.
 - `\Wizaplace\Favorite\FavoriteService::removeDeclinationToUserFavorites` is now taking a `string $declinationId` parameter instead of a `int $declinationId` one.
 - `\Wizaplace\Order\Order::getStatus` is now returning an `\Wizaplace\Order\OrderStatus` instead of a string.
+- `\Wizaplace\User\UserService::updateUser` now takes a single `\Wizaplace\User\UpdateUserCommand` instead of multiple parameters
 
 ### New features