+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 
 ### Bugfixes
 
+- `\Wizaplace\Order\OrderService::getOrderReturns` now uses the right API endpoint (it was systematically throwing 404 errors before)
+
 ## 0.7.3
 
 ### New features