+++ CHANGELOG.md
@@ -8,6 +8,10 @@
 
 ### Bugfixes
 
+## 0.7.4
+
+### Bugfixes
+
 - `\Wizaplace\Order\OrderService::getOrderReturns` now uses the right API endpoint (it was systematically throwing 404 errors before)
 
 ## 0.7.3